Move codegen generation to integration-tests folder
