+++ package.json
@@ -100,7 +100,7 @@
     "test:type-benches": "find src -name '*.bench.ts' | xargs -n1 tsx",
     "test:debug": "node --inspect-brk node_modules/.bin/jest --config ./config/jest.config.ts --runInBand --testTimeout 99999 --logHeapUsage",
     "test:ci": "TEST_ENV=ci npm run test:coverage -- --logHeapUsage",
-    "test:codegen": "npm run build -w codegen && graphql-codegen --config ./tests.codegen.ts",
+    "test:codegen": "npm run build -w codegen && cd integration-tests/codegen && npm run codegen",
     "test:watch": "jest --config ./config/jest.config.ts --watch",
     "test:memory": "npm i && npm run build && cd scripts/memory && npm i && npm test",
     "test:coverage": "npm run coverage -- --ci --runInBand --reporters=default --reporters=jest-junit",
@@ -169,7 +169,6 @@
     "@babel/preset-env": "^7.26.9",
     "@changesets/changelog-github": "0.5.0",
     "@changesets/cli": "2.27.7",
-    "@graphql-codegen/cli": "^5.0.6",
     "@graphql-tools/schema": "10.0.33",
     "@jest/expect-utils": "30.2.0",
     "@jest/globals": "30.2.0",
